Initial work on using strategies instead of iter
The old way was basically forcing each way of generating to be an
instance and contain a lot of boilerplate. Now they are defined as
strategies of generating, and are spawned as a blocking process for each
incoming request.

This has several advantages (although some profiling needs to be done):

1. Much less boilterplate
2. Strategies are now much more free, as long as they generate data.
   You could perhaps even make one connecting two clients together!
3. You can now buffer ready data while sending. TODO: Make buffer size
   configurable.

Still a lot to be done, but still.