Use strategies instead of iter (#45)
Initial work on using strategies instead of iter

The old way was basically forcing each way of generating to be an
instance and contain a lot of boilerplate. Now they are defined as
strategies of generating, and are spawned as a blocking process for each
incoming request.

This has several advantages (although some profiling needs to be done):

1. Much less boilterplate
2. Strategies are now much more free, as long as they generate data.
   You could perhaps even make one connecting two clients together!
3. You can now buffer ready data while sending. TODO: Make buffer size
   configurable.

Disadvantages:

May increase memory usage (but depends on `chunk_buffer` setting), as well as wasting some CPU time when creating buffered messages that are never sent.